+++ Povox/src/Platform/OpenGL/OpenGLBuffer.h
@@ -18,7 +18,7 @@ namespace Povox {
 		virtual void Bind() const override;
 		virtual void Unbind() const override;
 
-		virtual void Submit(const void* data, size_t size) const override;
+		virtual void Submit(const void* data, uint32_t size) const override;
 
 		virtual const BufferLayout& GetLayout() const override { return m_Layout; }
 		virtual void SetLayout(const BufferLayout& layout) override { m_Layout = layout;  }
@@ -44,7 +44,7 @@ namespace Povox {
 		virtual void Bind() const override;
 		virtual void Unbind() const	override;
 
-		virtual void Submit(const void* data, size_t size) const override;
+		virtual void Submit(const void* data, uint32_t size) const override;
 
 		inline virtual uint32_t GetCount() const override { return m_Count; }
