+++ Povox/src/Povox/Renderer/Buffer.h
@@ -101,7 +101,7 @@ namespace Povox {
 		virtual void Bind() const = 0;
 		virtual void Unbind() const = 0;
 
-		virtual void Submit(const void* data, size_t size) const = 0;
+		virtual void Submit(const void* data, uint32_t size) const = 0;
 
 		virtual void SetLayout(const BufferLayout& layout) = 0;
 		virtual const BufferLayout& GetLayout() const = 0;
@@ -120,7 +120,7 @@ namespace Povox {
 		virtual void Bind() const = 0;
 		virtual void Unbind() const = 0;
 
-		virtual void Submit(const void* data, size_t size) const = 0;
+		virtual void Submit(const void* data, uint32_t size) const = 0;
 
 		virtual uint32_t GetCount() const = 0;
